feat: Initial Commit and Release
Baski has the ability to Build, Scan and Sign OpenStack iamge using the Kubernetes-sigs/Image-builder under the hood
It has been designed to make building image a little smoother whilst adding things like scanning into the mix.
Scan results will be uploaded to a designated S3 endpoint for later review.

There is also a server which can be deployed via a supplied Helm chart which will pull scan information from the supplied S3 repo to display in a user firendly format. This may later be modified to enable approval or rejection of CVEs making images public/private or tagging them as vulnerable as a result.